pyproject.toml
[project]
name = "yp-autotests-sprint-3"
version = "0.1.0"
package-mode = false
description = ""
authors = [
    {name = "PaulNero",email = "58290662+PaulNero@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "selenium (>=4.36.0,<5.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "allure-pytest (>=2.15.0,<3.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
